refactor(canon): keep diagnostic files within the source-length budget
The path restoration added lines to two files already over the 350-line
limit, which the source-length guard rejects. Fold the CLI's per-message
loop into restore_authored_paths_in_messages next to the rewrite it
calls, and move the dedup key, dedup_diagnostics, and its test into a
dedup submodule alongside the other diagnostics submodules.
