+++ crates/vize_canon/src/batch/executor/diagnostics.rs
@@ -5,8 +5,9 @@ use serde_json::Value;
 use super::super::{Diagnostic, OriginalPosition, VirtualFile, VirtualProject};
 use crate::corsa_client::LspDiagnostic;
 use crate::file_uri::file_uri_to_path;
-use vize_carton::{FxHashMap, FxHashSet, String};
+use vize_carton::{FxHashMap, String};
 
+mod dedup;
 mod keyof_indexed_assignment;
 mod line_index;
 mod module_resolution;
@@ -14,10 +15,12 @@ mod module_specifier;
 mod skip_rules;
 mod virtual_path_message;
 
+pub(super) use dedup::dedup_diagnostics;
 use line_index::LineIndex;
 pub(super) use module_resolution::relative_module_resolves_on_disk;
 pub(super) use skip_rules::{should_skip_diagnostic, should_skip_original_diagnostic};
-pub(super) use virtual_path_message::restore_authored_paths;
+use virtual_path_message::restore_authored_paths;
+pub(super) use virtual_path_message::restore_authored_paths_in_messages;
 
 pub(super) fn map_batch_diagnostics(
     results: Vec<(String, Vec<LspDiagnostic>)>,
@@ -39,42 +42,6 @@ pub(super) fn map_batch_diagnostics(
     }
 
     dedup_diagnostics(diagnostics)
-}
-
-/// Identity key for deduplicating diagnostics — (file, line, column, code,
-/// message). After source mapping, distinct virtual positions can collapse to
-/// the same original position: a template binding (e.g. an undefined name in an
-/// interpolation) is referenced more than once in the generated virtual TS —
-/// once by the normal template-expression statement and once by the dedicated
-/// "Undefined references from template" check — and every reference maps back
-/// to the same source span. Corsa then reports the same template error at each
-/// virtual position, which would otherwise surface multiple times (#1389).
-/// Severity is part of the key so a genuine error+hint pair on the same span is
-/// preserved.
-type DiagnosticKey = (std::path::PathBuf, u32, u32, Option<u32>, String, u8);
-
-fn diagnostic_key(diagnostic: &Diagnostic) -> DiagnosticKey {
-    (
-        diagnostic.file.clone(),
-        diagnostic.line,
-        diagnostic.column,
-        diagnostic.code,
-        diagnostic.message.clone(),
-        diagnostic.severity,
-    )
-}
-
-/// Drop exact-duplicate diagnostics while preserving first-seen order, keyed on
-/// (file, line, column, code, message, severity).
-pub(super) fn dedup_diagnostics(diagnostics: Vec<Diagnostic>) -> Vec<Diagnostic> {
-    let mut seen: FxHashSet<DiagnosticKey> = FxHashSet::default();
-    let mut deduped = Vec::with_capacity(diagnostics.len());
-    for diagnostic in diagnostics {
-        if seen.insert(diagnostic_key(&diagnostic)) {
-            deduped.push(diagnostic);
-        }
-    }
-    deduped
 }
 
 pub(super) struct DiagnosticMapper<'a> {
@@ -361,47 +328,6 @@ mod tests {
         );
         assert_eq!(mapped[0].file, app_path);
         assert_eq!(mapped[0].code, Some(2304));
-    }
-
-    /// Distinct diagnostics on the same span (different code or message, or a
-    /// genuine error+hint pair) must survive deduplication.
-    #[test]
-    fn dedup_preserves_distinct_diagnostics() {
-        use super::dedup_diagnostics;
-        use crate::batch::Diagnostic;
-
-        let base = Diagnostic {
-            file: PathBuf::from("/p/App.vue"),
-            line: 4,
-            column: 6,
-            message: "Cannot find name 'x'.".into(),
-            code: Some(2304),
-            severity: 1,
-            block_type: Some(SfcBlockType::Template),
-        };
-        let duplicate = base.clone();
-        let different_code = Diagnostic {
-            code: Some(2322),
-            ..base.clone()
-        };
-        let different_message = Diagnostic {
-            message: "Cannot find name 'y'.".into(),
-            ..base.clone()
-        };
-        let different_severity = Diagnostic {
-            severity: 4,
-            ..base.clone()
-        };
-
-        let deduped = dedup_diagnostics(vec![
-            base.clone(),
-            duplicate,
-            different_code,
-            different_message,
-            different_severity,
-        ]);
-
-        assert_eq!(deduped.len(), 4, "{deduped:#?}");
     }
 
     #[test]
